Move expand-function tests to their own test file
